+++ moa/src/main/java/moa/classifiers/trees/HoeffdingOptionTree.java
@@ -205,11 +205,11 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
             this.observedClassDistribution = new DoubleVector(classObservations);
         }
 
-        public int calcByteSize() {
-            return (int) (SizeOf.sizeOf(this) + SizeOf.fullSizeOf(this.observedClassDistribution));
+        public long calcByteSize() {
+            return  SizeOf.sizeOf(this) + SizeOf.fullSizeOf(this.observedClassDistribution);
         }
 
-        public int calcByteSizeIncludingSubtree() {
+        public long calcByteSizeIncludingSubtree() {
             return calcByteSize();
         }
 
@@ -290,14 +290,14 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
         protected AutoExpandVector<Node> children = new AutoExpandVector<Node>();
 
         @Override
-        public int calcByteSize() {
+        public long calcByteSize() {
             return super.calcByteSize()
-                    + (int) (SizeOf.sizeOf(this.children) + SizeOf.fullSizeOf(this.splitTest));
+                    + SizeOf.sizeOf(this.children) + SizeOf.fullSizeOf(this.splitTest);
         }
 
         @Override
-        public int calcByteSizeIncludingSubtree() {
-            int byteSize = calcByteSize();
+        public long calcByteSizeIncludingSubtree() {
+            long byteSize = calcByteSize();
             for (Node child : this.children) {
                 if (child != null) {
                     byteSize += child.calcByteSizeIncludingSubtree();
@@ -539,9 +539,9 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
         }
 
         @Override
-        public int calcByteSize() {
+        public long calcByteSize() {
             return super.calcByteSize()
-                    + (int) (SizeOf.fullSizeOf(this.attributeObservers));
+                    + SizeOf.fullSizeOf(this.attributeObservers);
         }
 
         @Override
@@ -617,8 +617,8 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
 
     protected int maxPredictionPaths;
 
-    public int calcByteSize() {
-        int size = (int) SizeOf.sizeOf(this);
+    public long calcByteSize() {
+        long size =  SizeOf.sizeOf(this);
         if (this.treeRoot != null) {
             size += this.treeRoot.calcByteSizeIncludingSubtree();
         }
@@ -626,7 +626,7 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
     }
 
     @Override
-    public int measureByteSize() {
+    public long measureByteSize() {
         return calcByteSize();
     }
 
@@ -1040,7 +1040,7 @@ public class HoeffdingOptionTree extends AbstractClassifier implements MultiClas
             this.inactiveLeafByteSizeEstimate = (double) totalInactiveSize
                     / this.inactiveLeafNodeCount;
         }
-        int actualModelSize = this.measureByteSize();
+        long actualModelSize = this.measureByteSize();
         double estimatedModelSize = (this.activeLeafNodeCount
                 * this.activeLeafByteSizeEstimate + this.inactiveLeafNodeCount
                 * this.inactiveLeafByteSizeEstimate);
